tests/wpt/tests/web-animations/responsive/font-size-adjust.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id='container'>
    <div id='element'></div>
</div>
<script>

var container = document.getElementById('container');
var element = document.getElementById('element');

test(function() {
    var keyframes = [
        {fontSizeAdjust: 'inherit'},
        {fontSizeAdjust: '6'}
    ];

    container.style.fontSizeAdjust = 'none';
    var player = element.animate(keyframes, 20);

    player.pause();
    player.currentTime = 5;
    assert_equals(getComputedStyle(element).fontSizeAdjust, 'none');

    container.style.fontSizeAdjust = '2';
    assert_equals(getComputedStyle(element).fontSizeAdjust, '3');

    player.currentTime = 10;
    container.style.fontSizeAdjust = '8';
    assert_equals(getComputedStyle(element).fontSizeAdjust, '7');
}, 'font-size-adjust responsive to inherited font-size-adjust changes');
</script>
