LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/font-size-adjust-expected.txt


FAIL font-size-adjust responsive to inherited font-size-adjust changes assert_equals: expected (string) "none" but got (undefined) undefined
